fix: correct Chart.js SRI hash and cast daily amounts to float
Agent-Logs-Url: https://github.com/Tony201011/ListingSite/sessions/030d0e6f-e754-46ee-b1c4-f637242c6800

+++ app/Actions/Subscription/GetPurchaseHistory.php
@@ -75,7 +75,7 @@ class GetPurchaseHistory
         while ($current->lte($endDate)) {
             $dayKey = $current->format('Y-m-d');
             $chartLabels[] = $current->format('d M');
-            $chartAmounts[] = $dailyTotals[$dayKey]->total ?? 0;
+            $chartAmounts[] = (float) ($dailyTotals[$dayKey]->total ?? 0);
             $chartCounts[] = (int) ($dailyTotals[$dayKey]->count ?? 0);
             $current->addDay();
         }
